Support alternate serial port for debugging
This adds the ability to attach a TTL serial interface on JOY2.

I was going to use this to output debug logs without interfering with
the NABU interface, but have decided to take a different approach.
It's still useful to have these pins defined, though. It should have
no effect if not used.